+++ common-libs/user-management/CHANGELOG.md
@@ -1,5 +1,8 @@
-# release 2.0.1 (2022-11-24)
-* add security fix for signIn passwordless flow 
+# release 2.2.0 (2022-11-29)
+* add security fix for signIn passwordless flow
+# release 2.1.0 (2022-11-22)
+* Added `signUpWithProfile` and `logInWithProfile` methods.
+* Added `TrueCallerService` class for handling `Truecaller flow` specific operations.
 # release 2.0.0 (2022-10-20)
 BREAKING CHANGES:
 * username building strategy was changed to use uuidV4 as username